addTest and SQL Scripts

+++ individualTask/src/main/java/org/teamscore/individualTask/services/CategoryService.java
@@ -3,6 +3,9 @@ package org.teamscore.individualTask.services;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.teamscore.individualTask.exceptions.CannotDeleteEntityException;
+import org.teamscore.individualTask.models.DTO.entity.CategoryDTO;
+import org.teamscore.individualTask.models.DTO.entity.createDTO.CreateCategoryDTO;
 import org.teamscore.individualTask.models.entity.Category;
 import org.teamscore.individualTask.repositories.CategoryRepository;
 
@@ -14,37 +17,47 @@ public class CategoryService {
 
     @Autowired
     private CategoryRepository categoryRepository;
+    @Autowired
+    private ConverterDTOService converterDTOService;
 
-    public Category createCategory(Category category) {
-        category = categoryRepository.save(category);
-        return category;
+    public CategoryDTO createCategory(CreateCategoryDTO categoryDTO) {
+        Category category = new Category(
+                categoryDTO.getName(),
+                categoryDTO.getColor(),
+                categoryDTO.getDescription()
+        );
+        return converterDTOService.categoryConvert(categoryRepository.save(category));
     }
 
-    public Category updateCategory(Category category) {
-        var oldCategory = categoryRepository.findById(category.getId()).orElseThrow();
-        oldCategory.setColor(category.getColor());
-        oldCategory.setName(category.getName());
-        oldCategory.setDescription(category.getDescription());
-        return categoryRepository.save(oldCategory);
+    public CategoryDTO updateCategory(CategoryDTO category) {
+        var oldCategory = categoryRepository.findById(category.getId());
+        if (oldCategory.isEmpty())
+            return null;
+        var oldCategoryPresent = oldCategory.get();
+        oldCategoryPresent.setColor(category.getColor());
+        oldCategoryPresent.setName(category.getName());
+        oldCategoryPresent.setDescription(category.getDescription());
+        return converterDTOService.categoryConvert(categoryRepository.save(oldCategoryPresent));
     }
 
     public void deleteCategory(Long id) {
-        categoryRepository.deleteById(id);
+        var category = categoryRepository.findById(id).orElse(null);
+        deleteCategory(category);
     }
 
     public void deleteCategory(Category category) {
-        categoryRepository.delete(category);
+        if (category != null) {
+            if (!category.getCosts().isEmpty())
+                throw new CannotDeleteEntityException(String.format("Невозможно удалить %s с ID %d, так как он используется", category.getName(), category.getId()));
+            categoryRepository.delete(category);
+        }
     }
 
-    public List<Category> getAllCategory(Pageable pageable) {
-        return categoryRepository.findAll(pageable);
+    public List<CategoryDTO> getAllCategory(Pageable pageable) {
+        return categoryRepository.findAll(pageable).stream().map(cat -> converterDTOService.categoryConvert(cat)).toList();
     }
 
-    public Category getCategoryByName(String name) {
-        return categoryRepository.findByName(name).orElseThrow();
+    public CategoryDTO getCategoryByName(String name) {
+        return converterDTOService.categoryConvert(categoryRepository.findByName(name).orElse(null));
     }
-
-//    public List<Category> getAllCategoryWithCostByPeriod(LocalDate from, LocalDate to){
-//        return categoryRepository
-//    }
 }
